Create Day 3
This code first checks if num is equal to 0, in which case it returns 0 since the sum of its digits is 0. If num is not 0, it uses a formula to calculate the sum of its digits. The formula is based on the fact that the sum of the digits of a number num is equal to num % 9 if num is not divisible by 9, and 9 otherwise. Therefore, if num is divisible by 9, the method returns 9. Otherwise, it returns num % 9.