Create Day 4
We start with even and odd both set to 0. Then we loop through each bit in the binary representation of n, starting from the least significant bit (i.e., the rightmost bit). For each bit, we check whether it is 1 by using the bitwise AND operator (&) with 1. If the bit is 1, we increment either the even or odd counter depending on whether the bit's position is even or odd.

After processing each bit, we shift n to the right by 1 bit position using the right shift operator (>>). This moves the next bit into the least significant position so we can check it in the next iteration of the loop. We repeat this process until n becomes 0.

Finally, we return an array containing the values of even and odd counters.